requirements.txt
asgiref==3.8.1
convertdate==2.4.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
Faker==37.3.0
gunicorn==23.0.0
lunardate==0.2.2
packaging==25.0
psycopg2==2.9.10
pyluach==2.2.0
PyMeeus==0.5.12
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
workalendar==17.0.0
